requirements.txt
RPi.GPIO==0.6.2
ipython
